pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "drawcustom"
version = "0.1.0"
description = "Pure image rendering from drawing instructions for e-paper displays"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "g4bri3lDev", email = "admin@g4bri3l.de"}
]
keywords = ["e-paper", "eink", "image", "rendering", "drawcustom"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "Pillow>=10.0.0",
    "qrcode[pil]>=7.4.2",
    "aiohttp>=3.8.0",
    "resize-image>=0.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "mypy>=1.19.1",
    "ruff>=0.14.10",
]

[project.urls]
Homepage = "https://github.com/g4bri3lDev/drawcustom"
Repository = "https://github.com/g4bri3lDev/drawcustom"
Issues = "https://github.com/g4bri3lDev/drawcustom/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/drawcustom"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]